+++ src/database/models.py
@@ -200,3 +200,16 @@ class MarketItemModel(BaseModel):
 
     def __init__(self, **kwargs) -> None:
         super().__init__(**kwargs)
+
+
+class DailyGiftModel(BaseModel):
+    _id = Field(ObjectId)
+    owner = Field(ObjectId)
+    last_claimed_at = Field(datetime, nullable=True)
+    next_claimable_at = Field(datetime, default=datetime.utcnow() + timedelta(days=1))
+    is_claimed = Field(bool, default=False)
+    items = Field(list)
+    streak = Field(int, default=0)
+
+    def __init__(self, **kwargs) -> None:
+        super().__init__(**kwargs)