+++ src/main.py
@@ -1,6 +1,7 @@
 from threading import Thread
-import argparse
+import importlib
 
+import argparse
 from telebot.types import BotCommand
 
 import bot as _  # noqa: F401
@@ -50,16 +51,36 @@ def start_threads():
         thread.start()
 
 
+def reload_modules():
+    import base
+    import bot
+    import database
+    import helpers
+    import middlewares
+    import threads
+
+    modules = [base, bot, database, helpers, middlewares, threads]
+
+    for module in modules:
+        importlib.reload(module)
+
+
 def main(args):
     logger.info("Бот включен")
 
     if args.debug:
+        import config
+
+        config.DEBUG = True
+        config.telebot.logger.setLevel(10)
+        config.logger.setLevel(10)
+        reload_modules()
         logger.warning("Бот работает в режиме DEBUG")
 
     configure_bot_commands()
 
+    init_middlewares()
     if not args.without_threads:
-        init_middlewares()
         start_threads()
 
     bot.infinity_polling(timeout=500, skip_pending=True)